17.5.22

+++ work 2/Cal.h
@@ -1,5 +1,5 @@
 #define pi 3.14
-float CalCircleArea(int n)   //用来算圆的面积 
+float CalCircleArea(float n)   //用来算圆的面积 
 {
 	float area;
 	area=pi*n*n; //n为半径，pi假定为3.14
